requirements.txt
python-dotenv==1.0.0

# Streamlit release notes: https://docs.streamlit.io/develop/quick-reference/release-notes
streamlit==1.39.0
pandas==2.2.3 # optional, used for Streamlit tables
requests==2.31.0  # for making HTTP requests

# sqlalchemy==2.0.35
# psycopg2-binary==2.9.9  # to connect to postgres

# openai==1.30.1
anthropic==0.42.0
# google-cloud-aiplatform==1.42.1
tiktoken==0.8.0

slack_sdk==3.34.0 # for sending messages as a slack bot